requirements.txt
pandas>=1.5.0
simpy>=4.1.1
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
jupyter>=1.0.0

